backend/requirements.txt
aiosqlite==0.22.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgi-lifespan==2.1.0
asyncmy==0.2.11
bcrypt==5.0.0
certifi==2025.11.12
cffi==2.0.0
click==8.3.1
cryptography==46.0.3
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.127.0
fastapi-users==15.0.3
fastapi-users-db-sqlalchemy==7.0.0
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
jinja2==3.1.6
makefun==1.16.0
markupsafe==3.0.3
packaging==25.0
pluggy==1.6.0
pwdlib==0.3.0
pycparser==2.23
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.12.0
pygments==2.19.2
pyjwt==2.10.1
pytest==9.0.2
pytest-asyncio==1.3.0
python-dotenv==1.2.1
python-multipart==0.0.21
sniffio==1.3.1
sqladmin==0.22.0
sqlalchemy==2.0.45
sqlmodel==0.0.29
starlette==0.50.0
typing-extensions==4.15.0
typing-inspection==0.4.2
uvicorn==0.40.0
wtforms==3.1.2